% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abort_api_error}
\alias{abort_api_error}
\title{Abort with OMOPHub API Error}
\usage{
abort_api_error(status, message, endpoint, call = rlang::caller_env())
}
\arguments{
\item{status}{HTTP status code.}

\item{message}{Error message.}

\item{endpoint}{API endpoint that failed.}

\item{call}{The calling environment.}
}
\description{
Abort with OMOPHub API Error
}
\keyword{internal}
