% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abort_auth_error}
\alias{abort_auth_error}
\title{Abort with Authentication Error}
\usage{
abort_auth_error(message, call = rlang::caller_env())
}
\arguments{
\item{message}{Error message.}

\item{call}{The calling environment.}
}
\description{
Abort with Authentication Error
}
\keyword{internal}
