% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abort_rate_limit}
\alias{abort_rate_limit}
\title{Abort with Rate Limit Error}
\usage{
abort_rate_limit(retry_after = NULL, call = rlang::caller_env())
}
\arguments{
\item{retry_after}{Seconds until rate limit resets.}

\item{call}{The calling environment.}
}
\description{
Abort with Rate Limit Error
}
\keyword{internal}
