% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abort_validation}
\alias{abort_validation}
\title{Abort with Validation Error}
\usage{
abort_validation(message, arg = NULL, call = rlang::caller_env())
}
\arguments{
\item{message}{Error message.}

\item{arg}{The argument that failed validation.}

\item{call}{The calling environment.}
}
\description{
Abort with Validation Error
}
\keyword{internal}
