% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bool_to_str}
\alias{bool_to_str}
\title{Convert Boolean to API String}
\usage{
bool_to_str(x)
}
\arguments{
\item{x}{A logical value.}
}
\value{
"true" or "false" string, or NULL if x is NULL.
}
\description{
Convert Boolean to API String
}
\keyword{internal}
