% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir.R
\name{compact_list}
\alias{compact_list}
\title{Helper to remove NULL entries from a named list.}
\usage{
compact_list(...)
}
\arguments{
\item{...}{Named arguments.}
}
\value{
A list with NULL entries removed.
}
\description{
Helper to remove NULL entries from a named list.
}
\keyword{internal}
