% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_error_message}
\alias{extract_error_message}
\title{Extract Error Message from API Response}
\usage{
extract_error_message(resp)
}
\arguments{
\item{resp}{An httr2 response object.}
}
\value{
A character string with the error message.
}
\description{
Extract Error Message from API Response
}
\keyword{internal}
