% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir.R
\name{fhir_resolve}
\alias{fhir_resolve}
\title{Resolve a FHIR Coding to an OMOP standard concept}
\usage{
fhir_resolve(client, ...)
}
\arguments{
\item{client}{An \link{OMOPHubClient} instance.}

\item{...}{Arguments passed to \code{FhirResource$resolve()}.}
}
\value{
A list with \code{input} and \code{resolution}.
}
\description{
Standalone wrapper for \code{client$fhir$resolve()}. See
\link{FhirResource} for full parameter documentation.
}
\examples{
\dontrun{
client <- OMOPHubClient$new(api_key = Sys.getenv("OMOPHUB_API_KEY"))
fhir_resolve(
  client,
  system = "http://snomed.info/sct",
  code = "44054006",
  resource_type = "Condition"
)
}
}
\seealso{
\link{FhirResource}, \code{\link[=fhir_resolve_batch]{fhir_resolve_batch()}},
\code{\link[=fhir_resolve_codeable_concept]{fhir_resolve_codeable_concept()}}
}
