% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir.R
\name{fhir_resolve_batch}
\alias{fhir_resolve_batch}
\title{Batch-resolve FHIR Codings}
\usage{
fhir_resolve_batch(client, ...)
}
\arguments{
\item{client}{An \link{OMOPHubClient} instance.}

\item{...}{Arguments passed to \code{FhirResource$resolve_batch()}.}
}
\value{
See \code{FhirResource$resolve_batch()}.
}
\description{
Standalone wrapper for \code{client$fhir$resolve_batch()}. When
\code{as_tibble = TRUE}, the result is a flat \code{tibble} suitable for
\code{dplyr}/\code{tidyr} pipelines.
}
\examples{
\dontrun{
client <- OMOPHubClient$new(api_key = Sys.getenv("OMOPHUB_API_KEY"))
tbl <- fhir_resolve_batch(
  client,
  codings = list(
    list(system = "http://snomed.info/sct", code = "44054006"),
    list(system = "http://loinc.org", code = "2339-0")
  ),
  as_tibble = TRUE
)
dplyr::filter(tbl, status == "resolved")
}
}
\seealso{
\link{FhirResource}, \code{\link[=fhir_resolve]{fhir_resolve()}},
\code{\link[=fhir_resolve_codeable_concept]{fhir_resolve_codeable_concept()}}
}
