% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir.R
\name{fhir_resolve_codeable_concept}
\alias{fhir_resolve_codeable_concept}
\title{Resolve a FHIR CodeableConcept with vocabulary preference}
\usage{
fhir_resolve_codeable_concept(client, ...)
}
\arguments{
\item{client}{An \link{OMOPHubClient} instance.}

\item{...}{Arguments passed to \code{FhirResource$resolve_codeable_concept()}.}
}
\value{
See \code{FhirResource$resolve_codeable_concept()}.
}
\description{
Standalone wrapper for \code{client$fhir$resolve_codeable_concept()}.
}
\examples{
\dontrun{
client <- OMOPHubClient$new(api_key = Sys.getenv("OMOPHUB_API_KEY"))
fhir_resolve_codeable_concept(
  client,
  coding = list(
    list(system = "http://snomed.info/sct", code = "44054006"),
    list(system = "http://hl7.org/fhir/sid/icd-10-cm", code = "E11.9")
  ),
  resource_type = "Condition"
)
}
}
\seealso{
\link{FhirResource}, \code{\link[=fhir_resolve]{fhir_resolve()}}, \code{\link[=fhir_resolve_batch]{fhir_resolve_batch()}}
}
