% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_transient_error}
\alias{is_transient_error}
\title{Check if Error is Transient (Retryable)}
\usage{
is_transient_error(resp)
}
\arguments{
\item{resp}{An httr2 response object.}
}
\value{
\code{TRUE} if the error is transient and should be retried.
}
\description{
Check if Error is Transient (Retryable)
}
\keyword{internal}
