% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_params}
\alias{join_params}
\title{Join List Elements for Query Parameter}
\usage{
join_params(x, sep = ",")
}
\arguments{
\item{x}{A character vector.}

\item{sep}{Separator (default comma).}
}
\value{
A single comma-separated string, or NULL if x is NULL/empty.
}
\description{
Join List Elements for Query Parameter
}
\keyword{internal}
