% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{new_omophub_hierarchy}
\alias{new_omophub_hierarchy}
\title{Create OMOPHub Hierarchy Result}
\usage{
new_omophub_hierarchy(data, meta = list(), type = "hierarchy")
}
\arguments{
\item{data}{Hierarchy data (ancestors or descendants).}

\item{meta}{Metadata including summary.}

\item{type}{Type of hierarchy ("ancestors" or "descendants").}
}
\value{
An \code{omophub_hierarchy} S3 object.
}
\description{
Create OMOPHub Hierarchy Result
}
\keyword{internal}
