% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{new_omophub_mappings}
\alias{new_omophub_mappings}
\title{Create OMOPHub Mappings Result}
\usage{
new_omophub_mappings(data, meta = list())
}
\arguments{
\item{data}{Mapping data.}

\item{meta}{Metadata.}
}
\value{
An \code{omophub_mappings} S3 object.
}
\description{
Create OMOPHub Mappings Result
}
\keyword{internal}
