% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{new_omophub_relationships}
\alias{new_omophub_relationships}
\title{Create OMOPHub Relationships Result}
\usage{
new_omophub_relationships(data, meta = list())
}
\arguments{
\item{data}{Relationship data.}

\item{meta}{Metadata.}
}
\value{
An \code{omophub_relationships} S3 object.
}
\description{
Create OMOPHub Relationships Result
}
\keyword{internal}
