% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{new_omophub_vocabularies}
\alias{new_omophub_vocabularies}
\title{Create OMOPHub Vocabularies Result}
\usage{
new_omophub_vocabularies(data, meta = list())
}
\arguments{
\item{data}{Vocabulary data.}

\item{meta}{Pagination metadata.}
}
\value{
An \code{omophub_vocabularies} S3 object.
}
\description{
Create OMOPHub Vocabularies Result
}
\keyword{internal}
