% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{paginate_lazy}
\alias{paginate_lazy}
\title{Create Lazy Pagination Iterator}
\usage{
paginate_lazy(fetch_fn, page_size = 100)
}
\arguments{
\item{fetch_fn}{A function that takes \code{page} and \code{size} arguments.}

\item{page_size}{Number of items per page. Default 100.}
}
\value{
An iterator object with \code{next_page()} and \code{has_more()} methods.
}
\description{
Creates an iterator that fetches pages on demand.
}
\keyword{Check}
\keyword{Get}
\keyword{Reset}
\keyword{are}
\keyword{available}
\keyword{beginning}
\keyword{current}
\keyword{if}
\keyword{internal}
\keyword{iterator}
\keyword{more}
\keyword{next}
\keyword{number}
\keyword{of}
\keyword{page}
\keyword{pages}
\keyword{results}
\keyword{to}
