% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_concept_id}
\alias{validate_concept_id}
\title{Validate Concept ID}
\usage{
validate_concept_id(concept_id, arg = "concept_id", call = rlang::caller_env())
}
\arguments{
\item{concept_id}{A concept ID to validate.}

\item{arg}{Argument name for error messages.}

\item{call}{The calling environment.}
}
\description{
Validate Concept ID
}
\keyword{internal}
