% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_pagination}
\alias{validate_pagination}
\title{Validate Page Parameters}
\usage{
validate_pagination(
  page,
  page_size,
  max_page_size = 1000,
  call = rlang::caller_env()
)
}
\arguments{
\item{page}{Page number.}

\item{page_size}{Page size.}

\item{max_page_size}{Maximum allowed page size.}

\item{call}{The calling environment.}
}
\description{
Validate Page Parameters
}
\keyword{internal}
